+++ src/main/java/abstractNewspaper/PhysicalNewspaperSubscription.java
@@ -6,7 +6,7 @@ public class PhysicalNewspaperSubscription extends NewspaperSubscription {
     public void setSubscriberAddress(String subscriberAddress) {
         int check = 0;
         for (int i = 0; i < subscriberAddress.length(); i++) {
-            int ascii = (int) subscriberAddress.charAt(i);
+            int ascii = subscriberAddress.charAt(i);
             if (ascii >= 48 && ascii <= 57) {
                 check = 1;
                 break;
